requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
dj-rest-auth @ git+https://github.com/Juunsik/dj-rest-auth.git@4fd3f5f0cb9eb336ba4a209f5c5ce6936c320068
Django==4.2
django-allauth==0.62.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Faker==25.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
openai==1.30.1
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
toposort==1.10
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
